Create .gitattributes to fix web-ext signing on Windows
Mozilla has a hard limit of 4 MB maximum on files to be signed in extensions.

The cadmium-playercore-1080p.js file is already very close to that limit (3.94 MB).

When git clones the repository to a Windows client, it changes all of the LF to CRLF in each file in the repository.  This is usually no problem for most files, but causes the cadmium js file to exceed 4 MB, thus preventing Windows clients from self-signing the extension and testing code updates.

This commit simply adds a .gitattributes file with a single line that tells git to refrain from replacing LF with CRLF for the cadmium js file.  If the cadmium js file is renamed in the future, this .gitattributes file will need to be updated.

I have tested this on my Windows client and it solves the problem.  Mac and Linux clients will be unaffected since they already use LF instead of CRLF.